+++ src/main/java/ally/ArchiveList.java
@@ -1,6 +1,6 @@
-package duke;
+package ally;
 
-import duke.task.Task;
+import ally.task.Task;
 
 import java.util.ArrayList;
 public class ArchiveList extends ArrayList<Task> {
