+++ src/main/java/ally/TaskList.java
@@ -1,6 +1,6 @@
-package duke;
+package ally;
 
-import duke.task.Task;
+import ally.task.Task;
 
 import java.util.ArrayList;
