+++ src/main/java/ally/parser/Parser.java
@@ -1,4 +1,4 @@
-package duke.parser;
+package ally.parser;
 
 /**
  * Abstraction of String parsing.
