+++ src/main/java/ally/task/Deadline.java
@@ -1,4 +1,4 @@
-package duke.task;
+package ally.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
