+++ src/main/java/ally/task/Todo.java
@@ -1,4 +1,4 @@
-package duke.task;
+package ally.task;
 
 /**
  * Abstraction of Todo Tasks.
