+++ src/test/java/ally/TaskListTest.java
@@ -1,6 +1,6 @@
-package duke;
+package ally;
 
-import duke.task.Task;
+import ally.task.Task;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
